requirements.txt
# GaiaNet and AI
#gaianet==1.0.0  # Assuming this is the GaiaNet library
#torch==2.0.1
#transformers==4.30.2
#sentence-transformers==2.2.2

fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
pydantic-settings==2.0.2
python-dotenv==0.19.0
sqlalchemy==1.4.23
requests==2.26.0
web3==5.24.0
transformers==4.11.3
torch==2.0.1  # Updated version
datasets==1.11.0
wandb==0.12.1
qdrant-client==0.7.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==3.2.0
aiohttp==3.7.4
tqdm==4.62.2
# GaiaNet dependencies will be installed separately